+++ codes/janken02.js
@@ -1,0 +1,14 @@
+// じゃんけんの勝ち負けの結果を表示する関数
+const jankenHandler = (event) => {
+  alert("あなたの勝ちです")
+}
+
+// イベントリスナの設定
+// グー・チョキ・パー ぞれぞれのボタンが押されたときに、
+// jankenHandler関数が呼ばれるように、登録する。
+const guuButton   = document.querySelector("#guu")
+const chokiButton = document.querySelector("#choki")
+const paaButton   = document.querySelector("#paa")
+guuButton.addEventListener("click", jankenHandler)
+chokiButton.addEventListener("click", jankenHandler)
+paaButton.addEventListener("click", jankenHandler)